requirements.txt
# Core dependencies
python-dotenv==1.2.1
tavily-python==0.7.13

# LangChain and Google Gemini
langchain==1.0.7
langchain-core==1.0.5
langchain-google-genai==3.0.3

# Google API dependencies (installed automatically with langchain-google-genai)
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
